switched to double for lmsensor values, implemented uid for sensors, implemented getAll

+++ src/shared/headers/i_sensor_reader.h
@@ -4,5 +4,5 @@
 class ISensorReader {
     public:
         virtual std::vector<std::string> getAll() = 0;
-        virtual int getValue(std::string sensorId) = 0;
+        virtual double getValue(std::string sensorId) = 0;
 };
